Updates to pedometer
Now reads true to steps rather than wrapping from 0 to 255 (8-bit int now 32-bit)

+++ PeripheralDeviceFiles/companionplus_device/companionplus_device.ino
@@ -46,7 +46,7 @@ int currentLoop = 0;
 int flopX = 0;
 int flopY = 0;
 int flopZ = 0;
-uint8_t totalSteps = 0;
+uint32_t totalSteps = 0;
 float accelerationSense = 25;
 
 // Piezo buzzer tone
